+++ facade/src/main/java/com/timevale/forward/facade/api/request/BizDemandModifyReq.java
@@ -28,7 +28,7 @@ public class BizDemandModifyReq extends BaseReq {
     private Long deptId;
 
     @ApiModelProperty("优先级： 0-紧急，10-高，20-中，30低")
-    private Integer priority;
+    private Byte priority;
 
     @ApiModelProperty("产品线id")
     @JsonSerialize(using = ToStringSerializer.class)
@@ -44,10 +44,10 @@ public class BizDemandModifyReq extends BaseReq {
     private Boolean createCustomer;
 
     @ApiModelProperty("操作系统:0 XP, 1 Win7,2 Win8，3 Win10，4中标麒麟，5银河麒麟，6麒麟V10，7中科方德，8统信UOS")
-    private Integer os;
+    private Byte os;
 
     @ApiModelProperty("处理器: 0X86/X64, 1兆芯，2飞腾，3龙芯，4鲲鹏，5申威")
-    private Integer processor;
+    private Byte processor;
 
     @ApiModelProperty("需求描述")
     private String desc;
